+++ modules/ActCommon/deps.ts
@@ -1,0 +1,3 @@
+export * as rxjs from 'https://esm.sh/rxjs@7.8.1'
+export { act } from '../AtlasRenderer/mod.ts';
+