+++ modules/DryErase/deps.ts
@@ -1,7 +1,7 @@
 export * as m from '../Models/mod.ts';
 export * as rxjs from 'https://esm.sh/rxjs@7.8.1';
 export * as _ from "https://esm.sh/lodash-es@4.17.21";
-export { act } from '../AtlasRenderer/mod.ts';
+export { act, three } from '../AtlasRenderer/mod.ts';
 export * as hash from "https://esm.sh/fast-hash-code@2.1.0";
 export { nanoid } from "https://esm.sh/nanoid@4.0.2";
 
@@ -12,6 +12,8 @@ export * as channel from '../Data/ChannelCommon/mod.ts';
 export * as skia from '../Skia/mod.ts';
 export * as kayo from '../Kayo/mod.ts';
 export * as desk from '../Deskplane/mod.ts';
+export * as actCommon from '../ActCommon/mod.ts';
+export * as schedule from '../FrameScheduler/mod.ts';
 export * as boxParticle from '../BoxParticle/mod.ts';