+++ modules/Skia/deps.ts
@@ -1,4 +1,4 @@
 export { act } from '../AtlasRenderer/mod.ts';
 export * as rxjs from 'https://esm.sh/rxjs@7.8.1';
 
-export * as canvaskit from "https://esm.sh/canvaskit-wasm@0.38.2";
+export * as canvaskit from "https://esm.sh/canvaskit-wasm@0.39.1";